fix: More defensive code when trying to drop tables

+++ migration/2_jonah_upgrade_autoincrement.php
@@ -20,19 +20,16 @@ class JonahUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('jonah_channels', 'channel_id', 'autoincrementKey');
-        try {
+        if (in_array('jonah_channels_seq', $this->tables())) {
             $this->dropTable('jonah_channels_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('jonah_stories', 'story_id', 'autoincrementKey');
-        try {
+        if (in_array('jonah_stories_seq', $this->tables())) {
             $this->dropTable('jonah_stories_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('jonah_tags', 'tag_id', 'autoincrementKey');
-        try {
+        if (in_array('jonah_tags_seq', $this->tables())) {
             $this->dropTable('jonah_tags_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
